move collation code into `data.collate`.
