+++ chemprop/v2/data/datasets.py
@@ -1,6 +1,6 @@
 from dataclasses import dataclass, field
 from functools import cached_property
-from typing import NamedTuple
+from typing import NamedTuple, Protocol
 
 import numpy as np
 from rdkit import Chem
@@ -26,6 +26,15 @@ class Datum(NamedTuple):
     weight: float
     lt_mask: np.ndarray | None
     gt_mask: np.ndarray | None
+
+
+class MolGraphDatasetProto(Protocol):
+    def __getitem__(self, idx) -> Datum:
+        pass
+
+
+class MolGraphDataset(Dataset, MolGraphDatasetProto):
+    pass
 
 
 class _MolGraphDatasetMixin:
@@ -125,7 +134,7 @@ class _MolGraphDatasetMixin:
 
 
 @dataclass
-class MoleculeDataset(Dataset, _MolGraphDatasetMixin):
+class MoleculeDataset(MolGraphDataset, _MolGraphDatasetMixin):
     """A `MolgraphDataset` composed of `MoleculeDatapoint`s
 
     Parameters
@@ -271,7 +280,7 @@ class MoleculeDataset(Dataset, _MolGraphDatasetMixin):
 
 
 @dataclass
-class ReactionDataset(Dataset, _MolGraphDatasetMixin):
+class ReactionDataset(MolGraphDataset, _MolGraphDatasetMixin):
     """A :class:`ReactionDataset` composed of :class:`ReactionDatapoint`s"""
 
     data: list[ReactionDatapoint]
