+++ chemprop/v2/featurizers/__init__.py
@@ -1,6 +1,6 @@
 from .protos import *
 from .atom import AtomFeaturizer
 from .bond import BondFeaturizer
-from .molgraph import MolGraph, BatchMolGraph
+from .molgraph import MolGraph
 from .molecule import MoleculeMolGraphFeaturizer
 from .reaction import CGRFeaturizer, RxnMode
