+++ chemprop/v2/models/__init__.py
@@ -1,10 +1,2 @@
-from .modules import (
-    MessagePassingBlock,
-    MessagePassingBlockBase,
-    MulticomponentMessagePassing,
-    AtomMessageBlock,
-    BondMessageBlock,
-)
 from .model import MPNN
-from ..nn.loss import LossFunction
-from ..nn.metrics import Metric, MetricRegistry
+from .multi import MulticomponentMPNN
